Fixed One to many relation ship and added creation timestamps

+++ botigaCuadres/src/main/java/com/itacademy/botigaCuadres/repository/PaintingRepository.java
@@ -8,7 +8,7 @@ import com.itacademy.botigaCuadres.dto.PaintingResponseDto;
 import com.itacademy.botigaCuadres.dto.ShopResponseDto;
 
 @Repository
-public interface PaintingRepository extends JpaRepository<PaintingResponseDto, Integer> {
+public interface PaintingRepository extends JpaRepository<PaintingResponseDto, Long> {
 	
 	Iterable<PaintingResponseDto> findByShop(ShopResponseDto shop);
