+++ botigaCuadres/src/main/java/com/itacademy/botigaCuadres/repository/ShopRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import com.itacademy.botigaCuadres.dto.ShopResponseDto;
 
 @Repository
-public interface ShopRepository extends JpaRepository<ShopResponseDto,Integer> {
+public interface ShopRepository extends JpaRepository<ShopResponseDto,Long> {
